Extract `lib.user.member_request`, `.exc`, and dependees
