+++ pycroft/lib/user/__init__.py
@@ -4,7 +4,6 @@ from ._old import (
     encode_type2_user_id,
     decode_type2_user_id,
     check_user_id,
-    HostAliasExists,
     setup_ipv4_networking,
     store_user_sheet,
     get_user_sheet,
@@ -45,30 +44,38 @@ from ._old import (
     group_send_mail,
     send_member_request_merged_email,
     send_confirmation_email,
-    LoginTakenException,
-    EmailTakenException,
-    UserExistsInRoomException,
-    UserExistsException,
-    NoTenancyForRoomException,
-    MoveInDateInvalidException,
     get_similar_users_in_room,
     check_similar_user_in_room,
     get_user_by_swdd_person_id,
-    check_new_user_data,
-    check_new_user_data_unused,
-    create_member_request,
-    finish_member_request,
-    user_from_pre_member,
-    confirm_mail_address,
-    get_member_requests,
     get_name_from_first_last,
-    delete_member_request,
-    merge_member_request,
-    get_possible_existing_users_for_pre_member,
     get_user_by_id_or_login,
     send_password_reset_mail,
     change_password_from_token,
     find_similar_users,
     are_names_similar,
 )
+from .member_request import (
+    create_member_request,
+    finish_member_request,
+    user_from_pre_member,
+    get_member_requests,
+    delete_member_request,
+    merge_member_request,
+    get_possible_existing_users_for_pre_member,
+    check_new_user_data,
+    check_new_user_data_unused,
+)
 
+from .mail_confirmation import (
+    confirm_mail_address,
+)
+
+from .exc import (
+    HostAliasExists,
+    LoginTakenException,
+    EmailTakenException,
+    UserExistsInRoomException,
+    UserExistsException,
+    NoTenancyForRoomException,
+    MoveInDateInvalidException,
+)
